ai_chat/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
azure-core==1.39.0
azure-identity==1.25.3
azure-mgmt-compute==37.2.0
azure-mgmt-core==1.6.0
azure-mgmt-resource==25.0.0
azure-monitor-query==2.0.0
boto3==1.42.87
botocore==1.42.87
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
distro==1.9.0
Django==4.2.30
djangorestframework==3.16.1
exceptiongroup==1.3.1
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-compute==1.47.0
google-cloud-monitoring==2.30.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
jiter==0.13.0
jmespath==1.1.0
Markdown==3.9
msal==1.36.0
msal-extensions==1.3.1
openai==2.31.0
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
zipp==3.23.1
